Move xiaomi_aqara base entity to separate module
