+++ homeassistant/components/xiaomi_aqara/lock.py
@@ -9,8 +9,8 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.event import async_call_later
 
-from . import XiaomiDevice
 from .const import DOMAIN, GATEWAYS_KEY
+from .entity import XiaomiDevice
 
 FINGER_KEY = "fing_verified"
 PASSWORD_KEY = "psw_verified"
